fix integration test dataframe ordering
